refactor: move LoadingWidgetBuilder to game_widget directory
