+++ packages/flame/lib/src/game/game_widget/loading_widget_builder.dart
