+++ packages/flame/lib/widgets.dart
@@ -1,7 +1,7 @@
 export 'src/anchor.dart';
+export 'src/game/game_widget/loading_widget_builder.dart';
 export 'src/widgets/animation_widget.dart';
 export 'src/widgets/components_notifier_builder.dart';
-export 'src/widgets/loading_widget_builder.dart';
 export 'src/widgets/nine_tile_box.dart';
 export 'src/widgets/sprite_button.dart';
 export 'src/widgets/sprite_widget.dart';
